--- fdp/methods/nstx/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Tue Dec 01 16:24:57 2015
-
-@author: ktritz
-"""
-
