--- fdp/methods/nstx/equilibria/__init__.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Tue Dec 01 16:35:32 2015
-
-@author: ktritz
-"""
-
-from utilities import create_efit_objs as _preprocess
-
-__all__ = ['_preprocess']
