--- fdp/methods/nstx/mpts/__init__.py
@@ -1,4 +1,0 @@
-from plot import plot
-from change_units import change_units as _postprocess
-
-__all__ = ['_postprocess']
