--- fdp/methods/nstx/usxr/__init__.py
@@ -1,3 +1,0 @@
-from utilities import sub_offset as _postprocess
-
-__all__ = ['_postprocess']
