--- fdp/modules/nstx/equilibria/efit.xml
@@ -1,14 +1,0 @@
-<?xml version='1.0'?>
-<container name='[efit]'>
-    <axis name='time' units='s' mdsnode='\gtime'></axis>
-    <axis name='atime' units='s' mdsnode='\atime'></axis>
-    <axis name='R' units='m' axes='time' mdsnode='\R'></axis>
-    <axis name='z' units='m' axes='time' mdsnode='\z'></axis>
-    <axis name='psin' title='Psi_N' axes='time' mdsnode='\psin'></axis>
-    <signal name='qpsi' title='Q_psi' axes='time, psin' mdsnode='\qpsi'></signal>
-    <signal name='wmhd' title='Stored Energy' units='J' axes='time' mdsnode='\wmhd'></signal>
-    <signal name='psirz' title='Psi (R, z)' units='J' axes = 'time, R, z' mdsnode='\psirz'>
-        <defaults method='plot' type='contour' nlevels='50'></defaults>
-    </signal>
-    <node name='userid' mdsnode='\[efit]::userid'></node>
-</container>
