--- fdp/modules/nstx/mpts/mpts.xml
@@ -1,18 +1,0 @@
-<?xml version='1.0'?>
-<container name='mpts' mdstree='activespec' mdspath='\top.mpts.output_data.best' title='Multipoint Thomson Scattering' desc='test'>
-    <node name='comment' mdsnode='comment'></node>
-    <axis name='radius' units='cm' error='fit_r_width' mdsnode='fit_radii'></axis>
-    <axis name='time' units='s' mdsnode='ts_times'></axis>
-    <defaults method='plot' type='contourf'></defaults>
-    <signal name='valid' units='' axes='radius, time' mdsnode='valid' title='Valid flag'></signal>
-    <signal name='te' units='keV' axes='radius, time' error='fit_te_err' mdsnode='fit_te' title='Electron Temperature'>
-        <defaults method='plot' maxrange='0.95'></defaults>
-    </signal>
-    <signal name='ne' units='cm^-3' axes='radius, time' error='fit_ne_err' mdsnode='fit_ne' title='Electron Density'></signal>
-    <container name='spline'>
-        <axis name='radius' units='cm' mdsnode='spline_radii'></axis>
-        <signal name='valid' units='' axes='radius, time' mdsnode='svalid' title='Spline valid flag'></signal>
-        <signal name='te' units='keV' axes='radius, time' mdsnode='spline_te' title='Electron Temperature (spline fit)'></signal>
-        <signal name='ne' units='cm^-3' axes='radius, time' mdsnode='spline_ne' title='Electron Density (spline fit)'></signal>
-    </container>
-</container>
