--- fdp/modules/nstx/usxr/usxr.xml
@@ -1,15 +1,0 @@
-<?xml version='1.0'?>
-<container name='usxr' mdstree='usxr' title='USXR'>
-    <container name='vtop' title='Vertical Top'>
-        <axis name='time' units='s' mdsnode='dim_of(\usxr_vtop_00.raw)'></axis>
-        <signal name='vtop{}' title='Vertical Top Channel {}' units='DN' axes='time' mdsnode='\usxr_vtop_{}.raw' range='16'></signal>
-    </container>
-    <container name='hup' title='Horizontal Down'>
-        <axis name='time' units='s' mdsnode='dim_of(\usxr_hup_00.raw)'></axis>
-        <signal name='hup{}' title='Channel {}' units='DN' axes='time' mdsnode='\usxr_hup_{}.raw' range='16'></signal>
-    </container>
-    <container name='hdown' title='Horizontal Down'>
-        <axis name='time' units='s' mdsnode='dim_of(\usxr_hdown_00.raw)'></axis>
-        <signal name='hdown{}' title='Channel {}' units='DN' axes='time' mdsnode='\usxr_hdown_{}.raw' range='16'></signal>
-    </container>
-</container>
